NAVI/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
asyncio==3.4.3
numpy==1.24.3